[putperf] collect plasma store config, make it accessible from malloc, works on a mac

+++ cpp/src/plasma/malloc.cc
@@ -31,6 +31,7 @@
 #include <cstring>
 
 #include "plasma/common.h"
+#include "plasma/plasma.h"
 
 extern "C" {
 void* fake_mmap(size_t);
@@ -84,6 +85,7 @@ static ptrdiff_t pointer_distance(void const* pfrom, void const* pto) {
  * immediately unlinking it so we do not leave traces in the system. */
 int create_buffer(int64_t size) {
   int fd;
+  std::string file_template = plasma::plasma_config->directory;
 #ifdef _WIN32
   if (!CreateFileMapping(INVALID_HANDLE_VALUE, NULL, PAGE_READWRITE,
                          (DWORD)((uint64_t)size >> (CHAR_BIT * sizeof(DWORD))),
@@ -92,51 +94,48 @@ int create_buffer(int64_t size) {
   }
 #else
 #ifdef __linux__
-#define FILE_NAME "/mnt/hugepages/hugepagefile"
-  // if (hugetlbfs) then
-  constexpr char file_template[] = FILE_NAME;
-  // else
+  if (plasma::plasma_config->hugetlb_enabled) {
+    file_template += "/hugepagefile";
+  } else {
+    file_template += "/plasmaXXXXXX"; // template
+  }
   // constexpr char file_template[] = "/dev/shm/plasmaXXXXXX";
 #else
-  constexpr char file_template[] = "/tmp/plasmaXXXXXX";
+  // constexpr char file_template[] = "/tmp/plasmaXXXXXX";
+  file_template += "/plasmaXXXXXX";
 #endif
-  char file_name[32];
-  // TODO(atumanov): for hugetlbfs use hugetlbfs_unlinked_fd_for_size() from hugetlbfs.h
-  strncpy(file_name, file_template, 32);
-  //fd = mkstemp(file_name);
-  // create a file descriptor to a hugepage-based file.
-  fd = open(file_name, O_CREAT | O_RDWR, 0755);
+  std::vector<char> file_name(file_template.begin(), file_template.end());
+  file_name.push_back('\0');
+  if (plasma::plasma_config->hugetlb_enabled) {
+    // create a file descriptor to a hugepage-based file.
+    fd = open(&file_name[0], O_CREAT | O_RDWR, 0755);
+  } else {
+    fd = mkstemp(&file_name[0]);
+  }
   if (fd < 0) {
     perror("create_buffer: open failed");
-    ARROW_LOG(FATAL) << "create_buffer failed to open file " << file_name;
+    ARROW_LOG(FATAL) << "create_buffer failed to open file " << &file_name[0];
     return -1;
   }
-  //if (fd < 0) return -1;
+
   FILE* file = fdopen(fd, "a+");
   if (!file) {
     perror("create_buffer: fdopen failed");
     close(fd);
-    ARROW_LOG(FATAL) << "create_buffer: fdopen failed for " << file_name;
-    //exit(1);
+    ARROW_LOG(FATAL) << "create_buffer: fdopen failed for " << &file_name[0];
     return -1;
   }
-  if (unlink(file_name) != 0) {
+  if (unlink(&file_name[0]) != 0) {
     perror("create_buffer: failed to unlink file");
-#ifdef __linux__
-    ARROW_LOG(WARNING) << "unlink error";
-#else
     ARROW_LOG(FATAL) << "unlink error";
     return -1;
-#endif
   }
-  if (ftruncate(fd, (off_t)size) != 0) {
-    perror("create_buffer: failed to truncate file");
-#ifdef __linux__
-    ARROW_LOG(WARNING) << "create_buffer: ftruncate failed on hugetlbfs-backed file";
-#else
-    ARROW_LOG(FATAL) << "ftruncate error";
-    return -1;
-#endif
+  if (!plasma::plasma_config->hugetlb_enabled) {
+    if (ftruncate(fd, (off_t)size) != 0) {
+      perror("create_buffer: failed to truncate file");
+      ARROW_LOG(FATAL) << "ftruncate error";
+      return -1;
+    }
   }
 #endif
   return fd;
@@ -151,22 +150,18 @@ void* fake_mmap(size_t size) {
 
   int fd = create_buffer(size);
   ARROW_CHECK(fd >= 0) << "Failed to create buffer during mmap";
-#ifdef __linux__
-  //void *pointer = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
   void *pointer = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
   if (pointer == MAP_FAILED) {
-    std::cout << "mmap failed with error : " << std::strerror(errno) << std::endl;
+    ARROW_LOG(ERROR) << "mmap failed with error : " << std::strerror(errno);
     return pointer;
   }
   // Attempt to mlock the mmaped region of memory (best effort).
   int rv = mlock(pointer, size);
   if (rv != 0) {
-    std::cout << "mlock failed with error : " << std::strerror(errno) << std::endl;
+    ARROW_LOG(WARNING) << "mlock failed with error : " << std::strerror(errno);
   }
-  std::cout << "mlocking pointer " << pointer << " size " << size << " success " << rv << std::endl;
-#else
-  void *pointer = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
-#endif
+  ARROW_LOG(WARNING) << "mlocking pointer " << pointer << " size " << size
+                     << " success " << rv;
   memset(pointer, 0xff, size);
 
   /* Increase dlmalloc's allocation granularity directly. */
